pyproject.toml
[tool.poetry]
name = "hexgona-arch"
version = "0.1.0"
description = ""
authors = ["Jhon Valderrama <jhonvalderramaa7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
chromadb = "^0.5.5"
onnxruntime = "^1.18.1"
pydantic-settings = "^2.4.0"
protobuf = ">=3.20.0,<3.21.0"
openai = "^1.41.0"
pytest = "^8.3.3"
coverage = "^7.6.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
